requirements.txt
codeforcesapipy
flask
